requirements.txt
pydicom
requests
python-dotenv
requests